+++ src/spaceObjects/spaceObject.h
@@ -254,7 +254,7 @@ public:
     virtual bool canBeHackedBy(P<SpaceObject> other);
     virtual std::vector<std::pair<string, float> > getHackingTargets();
     virtual void hackFinished(P<SpaceObject> source, string target);
-    virtual void takeDamage(float damage_amount, DamageInfo info) {}
+    virtual void takeDamage(float damage_amount, DamageInfo info);
     virtual std::unordered_map<string, string> getGMInfo() { return std::unordered_map<string, string>(); }
     virtual string getExportLine() { return ""; }